Fence timeout is now an uint64_t (as should always have been).

+++ source/Renderer/GlRenderer/Src/Core/GlRenderingResources.cpp
@@ -16,7 +16,7 @@ namespace gl_renderer
 		m_fence = std::make_unique< Fence >( m_device, renderer::FenceCreateFlag::eSignaled );
 	}
 
-	bool RenderingResources::waitRecord( uint32_t timeout )
+	bool RenderingResources::waitRecord( uint64_t timeout )
 	{
 		bool res = m_fence->wait( timeout ) == renderer::WaitResult::eSuccess;
